pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcus Schiesser <mail@marcusschiesser.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
llama-index = "^0.9.19"
pypdf = "^3.17.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
alembic = "^1.13.1"
psycopg2 = {extras = ["binary"], version = "^2.9.9"}
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pgvector = "^0.2.4"
fsspec = "^2023.12.2"
s3fs = "^2023.12.2"
cachetools = "^5.3.2"


[tool.poetry.group.dev.dependencies]
fire = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "main:start"